Done with buttons handling, logger left

+++ src/CounterState.h
@@ -1,19 +1,21 @@
 #ifndef CounterState_h_
 #define CounterState_h_
 
+#include <Arduino.h>
+
 class CounterState {
   public:
     virtual void activate();
     virtual void nextStep();
     virtual void stop();
     virtual bool shouldStop();
-    virtual int stopReason();
+    virtual byte stopReason();
 
   protected:
     void log (const char* content);
 
   private:
-    int stageId;
+    byte stageId;
 };
 
 #endif
